Lots of implementation details for DirectX12.
- Added `DX12InputLayout`.
- Added `PipelineState` for DirectX12.
- Added `DX12BlendingState`.
- Added `DX12DepthStencilState`.
- Added `DX12VertexArray`.
- Added `DX12RasterizerState`.
- Began Adding `DX12ResourceTexture1D`.
- Several other minor changes.

+++ tau/TauEngine/src/dx/dx11/DX11RasterizerState.cpp
@@ -76,7 +76,7 @@ bool DX11RasterizerStateBuilder::processArgs(const RasterizerArgs& args, DXRaste
     dxArgs->desc.CullMode = dxCullMode(args.cullMode);
     dxArgs->desc.FillMode = dxFillMode(args.fillMode);
 
-    dxArgs->desc.DepthBias = args.depthBias;
+    dxArgs->desc.DepthBias = static_cast<INT>(args.depthBias);
     dxArgs->desc.DepthBiasClamp = args.depthBiasClamp;
     dxArgs->desc.SlopeScaledDepthBias = args.slopeScaledDepthBias;
     dxArgs->desc.DepthClipEnable = true;
